Codes changes during Integration tests
