+++ lom/int_test/config_files/globals.conf.json
@@ -1,16 +1,10 @@
 {
-
         "MAX_SEQ_TIMEOUT_SECS": 120,
-
         "MIN_PERIODIC_LOG_PERIOD_SECS": 1,
-
         "ENGINE_HB_INTERVAL_SECS": 10,
-
         "INITIAL_DETECTION_REPORTING_FREQ_IN_MINS": 5,
-
         "SUBSEQUENT_DETECTION_REPORTING_FREQ_IN_MINS": 60,
-
-        "INITIAL_DETECTION_REPORTING_MAX_COUNT": 12
-
-}
+        "INITIAL_DETECTION_REPORTING_MAX_COUNT": 12,
+        "PLUGIN_MIN_ERR_CNT_TO_SKIP_HEARTBEAT" : 3
+}
