+++ lom/src/engine/engine/engine_test.go
@@ -95,18 +95,18 @@ var globals_conf = `{"ENGINE_HB_INTERVAL_SECS" : 3 }`
 /*
  *  Actions.conf
  */
-var actions_conf = `{ "actions": [
-        { "name": "Detect-0" },
-        { "name": "Safety-chk-0", "Timeout": 1},
-        { "name": "Mitigate-0", "Timeout": 6},
-        { "name": "Detect-1" },
-        { "name": "Safety-chk-1", "Timeout": 7},
-        { "name": "Mitigate-1", "Timeout": 8},
-        { "name": "Detect-2" },
-        { "name": "Safety-chk-2", "Timeout": 1},
-        { "name": "Mitigate-2", "Timeout": 6},
-        { "name": "Disabled-0", "Disable": true}
-        ] }`
+var actions_conf = `{
+        "Detect-0" : { "name": "Detect-0" },
+        "Safety-chk-0" : { "name": "Safety-chk-0", "Timeout": 1},
+        "Mitigate-0" : { "name": "Mitigate-0", "Timeout": 6},
+        "Detect-1" : { "name": "Detect-1" },
+        "Safety-chk-1" : { "name": "Safety-chk-1", "Timeout": 7},
+        "Mitigate-1" : { "name": "Mitigate-1", "Timeout": 8},
+        "Detect-2" : { "name": "Detect-2" },
+        "Safety-chk-2" : { "name": "Safety-chk-2", "Timeout": 1},
+        "Mitigate-2" : { "name": "Mitigate-2", "Timeout": 6},
+        "Disabled-0" : { "name": "Disabled-0", "Disable": true}
+        }`
 
 var bindings_conf = `{ "bindings": [
     {
