+++ lom/src/pluginmgr/pluginmgr_common/pluginmgr_helper.go
@@ -243,24 +243,24 @@ func (plmgr *PluginManager) run() error {
             }
         case serverReq := <-serverReqChan: // requests from engine are received here and handled
             if plmgr.getShutdownStatus() {
-                LogInfo("In run() RecvServerRequest: Shutdown is active, so ignoring request: %v", serverReq)
+                LogInfo("In run() RecvServerRequest: Shutdown is active, so ignoring request: %s", lomipc.PrintServerRequest(serverReq, false))
                 if lomcommon.GetLoMRunMode() == lomcommon.LoMRunMode_Test {
                     return nil
                 }
             } else {
                 switch serverReq.ReqType {
                 case lomipc.TypeServerRequestAction:
-                    LogInfo("In run() RecvServerRequest : Received action request : %v", serverReq)
-                    actionReq, ok := serverReq.ReqData.(*lomipc.ActionRequestData)
+                    LogInfo("In run() RecvServerRequest : Received action request : %s", lomipc.PrintServerRequest(serverReq, false))
+                    actionReq, ok := serverReq.ReqData.(lomipc.ActionRequestData)
                     if !ok {
                         LogError("In run() RecvServerRequest : Error in parsing ActionRequestData for type : %v, data : %v",
                             serverReq.ReqType, serverReq.ReqData)
                     } else {
                         lomcommon.GetGoroutineTracker().Start("plg_mgr_Run_Action_"+actionReq.Action+"_"+lomcommon.GetUUID(),
-                            plmgr.handleRequest, actionReq)
+                            plmgr.handleRequest, &actionReq)
                     }
                 case lomipc.TypeServerRequestShutdown:
-                    LogInfo("In run() RecvServerRequest : Received shutdown request : %v", serverReq)
+                    LogInfo("In run() RecvServerRequest : Received shutdown request : %s", lomipc.PrintServerRequest(serverReq, false))
                     shutdownReq, ok := serverReq.ReqData.(*lomipc.ShutdownRequestData)
                     if !ok {
                         LogError("In run RecvServerRequest : Error in parsing ShutdownRequestData for type : %v, data : %v",
@@ -372,7 +372,7 @@ func (plmgr *PluginManager) handleRequest(actionReq *lomipc.ActionRequestData) e
             return
         }
 
-        LogDebug("In handleRequest(): Received response from plugin %v", respData.Action)
+        LogDebug("In handleRequest(): Received response from plugin %v, data : %s", respData.Action, lomipc.PrintActionResponseData(respData))
 
         if respData.Action != actionReq.Action {
             LogPanic("In handleRequest(): Invalid action name received. Got  %v, expected %v",
@@ -466,7 +466,7 @@ loop:
     for {
         select {
         case hbvalue := <-hbChan:
-            LogDebug("In handleRequest(): Received heartbeat from plugin %v", hbvalue.PluginName)
+            LogDebug("In handleRequest(): Received heartbeat from plugin %v, time : %v", hbvalue.PluginName, hbvalue.EpochTime)
             if hbvalue.PluginName != actionReq.Action {
                 LogPanic("In handleRequest(): Error, Received heartbeat from plugin %v, expected %v",
                     hbvalue.PluginName, actionReq.Action)
@@ -849,7 +849,7 @@ func SetupSignals() {
  * parse program command line arguments
  */
 
-func ParseArguments() {
+func ParseArguments() string {
     // Create a new flag set
     fs := flag.NewFlagSet("customFlags", flag.ExitOnError)
 
@@ -866,14 +866,17 @@ func ParseArguments() {
 
     if ProcIDFlag == "" {
         LogPanic("Exiting : Proc ID is not provided")
-        return
+        return ""
     }
 
     // assign to variables which can be accessed from process
     ProcID = ProcIDFlag
     lomcommon.SetLogLevel(syslog.Priority(syslogLevelFlag))
 
-    fmt.Printf("Program Arguments : proc ID : %s, Syslog Level : %d\n", ProcIDFlag, syslogLevelFlag)
+    msg := fmt.Sprintf("Program Arguments : proc ID : %s, Syslog Level : %d\n", ProcIDFlag, syslogLevelFlag)
+    fmt.Println(msg)
+
+    return msg
 }
 
 /*
@@ -934,10 +937,12 @@ func StartPluginManager(waittime time.Duration) error {
 func SetupPluginManager() error {
 
     //parse program arguments & assign values to program variables. Hree proc_X value is read
-    ParseArguments()
+    msg := ParseArguments()
 
     // setup application prefix for logging
     lomcommon.SetPrefix(APP_NAME_DEAULT + "_" + ProcID)
+
+    LogInfo(msg)
 
     //syslog level change from UNIX signals
     SetupSignals()
