+++ lom/src/pluginmgr/pluginmgr_main.go
@@ -4,7 +4,14 @@ import (
     "log/syslog"
     "lom/src/lib/lomcommon"
     "lom/src/pluginmgr/pluginmgr_common"
+
+    "lom/src/plugins/plugins_files"
+    "lom/src/plugins/plugins_files/sonic/plugin/linkcrc"
 )
+
+// TODO : Goutham : Temporary untill pernmant fix to include plugin files are found
+var t1 = plugins_files.NewGenericPluginDetection
+var t2 = linkcrc.NewLinkCRCDetectionPlugin
 
 /*
 * Main function for plugin manager
