--- lom/src/plugins/sonic/plugin_integration_tests/linkcrc/LINK_CRC_COUNTERS2.txt
@@ -1,6 +1,0 @@
-{
-  "SAI_PORT_STAT_IF_IN_ERRORS": 450,
-  "SAI_PORT_STAT_IF_IN_UCAST_PKTS": 222,
-  "SAI_PORT_STAT_IF_OUT_UCAST_PKTS": 2100,
-  "SAI_PORT_STAT_IF_OUT_ERRORS": 2
-}
